fix change input search channel event not change list mobile
